% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_features.R
\name{print.audit_features_result}
\alias{print.audit_features_result}
\title{Print audit results compactly}
\usage{
\method{print}{audit_features_result}(x, n = 10, ...)
}
\arguments{
\item{x}{An \code{audit_features_result} object.}

\item{n}{Maximum number of rows to print.}

\item{...}{Unused.}
}
\description{
Print audit results compactly
}
