% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobibeR.fr-package.R
\docType{package}
\name{pseudobibeR.fr}
\alias{pseudobibeR.fr-package}
\alias{pseudobibeR.fr}
\title{pseudobibeR.fr: French Biber Feature Extraction}
\description{
Tools for computing multidimensional register features in French,
complementing the English-focused \code{pseudobibeR} package. The current
development iteration focuses on building language resources (dictionaries,
word lists) and parsing helpers grounded in Universal Dependencies.
}
\author{
\strong{Maintainer}: David Brown \email{dwb2@andrew.cmu.edu} (\href{https://orcid.org/0000-0001-7745-6354}{ORCID})

}
\keyword{internal}
